[exotica_core] Fix float conversion, add as error

+++ exotica_core/include/exotica_core/tools.h
@@ -58,20 +58,20 @@ std_msgs::ColorRGBA RandomColor();
 inline std_msgs::ColorRGBA GetColor(double r, double g, double b, double a = 1.0)
 {
     std_msgs::ColorRGBA ret;
-    ret.r = r;
-    ret.g = g;
-    ret.b = b;
-    ret.a = a;
+    ret.r = static_cast<float>(r);
+    ret.g = static_cast<float>(g);
+    ret.b = static_cast<float>(b);
+    ret.a = static_cast<float>(a);
     return ret;
 }
 
 inline std_msgs::ColorRGBA GetColor(const Eigen::Vector4d& rgba)
 {
     std_msgs::ColorRGBA ret;
-    ret.r = rgba(0);
-    ret.g = rgba(1);
-    ret.b = rgba(2);
-    ret.a = rgba(3);
+    ret.r = static_cast<float>(rgba(0));
+    ret.g = static_cast<float>(rgba(1));
+    ret.b = static_cast<float>(rgba(2));
+    ret.a = static_cast<float>(rgba(3));
     return ret;
 }
